pyproject.toml
[tool.poetry]
name = "PyWhat"
version = "0.1.0"
description = "What is that thing?"
authors = ["Bee <github@skerritt.blog>"]
license = "GPLv3"

[tool.poetry.dependencies]
python = "^3.6"
click = "^7.1.2"
PyYAML = "^5.4.1"
name-that-hash = "^1.7.0"
langdetect = "^1.0.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
pywhat = "PyWhat.what:main"
what = "PyWhat.what:main"